WTA #44: Moved RecalculateLatestStreakService to use-cases package.
