+++ home/src/test/java/com/jacob/wakatimeapp/home/domain/usecases/CalculateCurrentStreakUCRobot.kt
@@ -5,7 +5,6 @@ import com.jacob.wakatimeapp.core.models.Error
 import com.jacob.wakatimeapp.core.models.Time
 import com.jacob.wakatimeapp.home.data.local.HomePageCache
 import com.jacob.wakatimeapp.home.domain.InstantProvider
-import com.jacob.wakatimeapp.home.domain.RecalculateLatestStreakService
 import com.jacob.wakatimeapp.home.domain.models.Last7DaysStats
 import com.jacob.wakatimeapp.home.domain.models.StreakRange
 import io.kotest.assertions.asClue
@@ -29,7 +28,7 @@ internal class CalculateCurrentStreakUCRobot {
     private var result: Either<Error, StreakRange>? = null
 
     private val mockCache: HomePageCache = mockk(relaxUnitFun = true)
-    private val mockRecalculateStreak: RecalculateLatestStreakService = mockk()
+    private val mockRecalculateStreak: RecalculateLatestStreakUC = mockk()
 
     fun buildUseCase() = apply {
         clearMocks(mockCache, mockRecalculateStreak)
@@ -42,7 +41,7 @@ internal class CalculateCurrentStreakUCRobot {
                 override fun now() = currentDayInstant
             },
             homePageCache = mockCache,
-            recalculateLatestStreakService = mockRecalculateStreak,
+            recalculateLatestStreakUC = mockRecalculateStreak,
         )
     }
